+++ packages/firebase_game_services/example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -1,0 +1,16 @@
+//
+//  Generated file. Do not edit.
+//
+
+import FlutterMacOS
+import Foundation
+
+import firebase_auth
+import firebase_core
+import firebase_game_services_apple
+
+func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  FLTFirebaseAuthPlugin.register(with: registry.registrar(forPlugin: "FLTFirebaseAuthPlugin"))
+  FLTFirebaseCorePlugin.register(with: registry.registrar(forPlugin: "FLTFirebaseCorePlugin"))
+  FirebaseGameServicesApplePlugin.register(with: registry.registrar(forPlugin: "FirebaseGameServicesApplePlugin"))
+}