+++ packages/firebase_game_services_google/android/src/main/kotlin/keew/ee/firebase_game_services_google/FirebaseGameServicesGooglePlugin.kt
@@ -3,64 +3,42 @@ package keew.ee.firebase_game_services_google
 import android.app.Activity
 import android.content.Context
 import android.content.Intent
-
 import android.util.Log
-import android.view.Gravity
-import androidx.annotation.NonNull
-
 import com.google.android.gms.auth.api.Auth
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.FirebaseAuthException
-import com.google.firebase.auth.PlayGamesAuthProvider
-
 import com.google.android.gms.auth.api.signin.GoogleSignIn
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount
 import com.google.android.gms.auth.api.signin.GoogleSignInClient
 import com.google.android.gms.auth.api.signin.GoogleSignInOptions
-import com.google.android.gms.auth.api.signin.GoogleSignInResult
-
 import com.google.android.gms.common.api.ApiException
 import com.google.android.gms.common.api.Status
-import com.google.android.gms.games.AchievementsClient
-import com.google.android.gms.games.Games
-import com.google.android.gms.games.LeaderboardsClient
-import com.google.android.gms.games.PlayersClient
-
-import com.google.android.gms.games.snapshot.Snapshot
-import com.google.android.gms.games.snapshot.SnapshotMetadata
-import com.google.android.gms.games.snapshot.SnapshotMetadataChange
-import com.google.android.gms.games.SnapshotsClient
-import com.google.android.gms.games.SnapshotsClient.DataOrConflict
+import com.google.android.gms.games.*
 import com.google.android.gms.tasks.Task
-
-import com.google.android.gms.drive.Drive
-import com.google.gson.Gson
-
+import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.auth.FirebaseAuthException
+import com.google.firebase.auth.PlayGamesAuthProvider
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.activity.ActivityAware
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
 import io.flutter.plugin.common.BinaryMessenger
-import io.flutter.plugin.common.MethodChannel.Result
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
+import io.flutter.plugin.common.MethodChannel.Result
 import io.flutter.plugin.common.PluginRegistry
-import java.lang.Exception
-import java.io.IOException
 
 private const val CHANNEL_NAME = "firebase_game_services"
 private const val RC_SIGN_IN = 9000
+private const val RC_ACHIEVEMENT_UI = 9003
+private const val RC_LEADERBOARD_UI = 9004
 
 class FirebaseGameServicesGooglePlugin(private var activity: Activity? = null) : FlutterPlugin,
     MethodChannel.MethodCallHandler, ActivityAware, PluginRegistry.ActivityResultListener {
 
     private var googleSignInClient: GoogleSignInClient? = null
-    private var snapshotsClient: SnapshotsClient? = null
     private var achievementClient: AchievementsClient? = null
     private var leaderboardsClient: LeaderboardsClient? = null
     private var activityPluginBinding: ActivityPluginBinding? = null
     private var channel: MethodChannel? = null
     private var pendingOperation: PendingOperation? = null
-    
     private lateinit var context: Context
 
     private var method: String? = null
@@ -70,7 +48,7 @@ class FirebaseGameServicesGooglePlugin(private var activity: Activity? = null) :
 
     companion object {
         @JvmStatic
-        fun getResourceFromContext(@NonNull context: Context, resName: String): String {
+        fun getResourceFromContext(context: Context, resName: String): String {
             val stringRes = context.resources.getIdentifier(resName, "string", context.packageName)
             if (stringRes == 0) {
                 throw IllegalArgumentException(
@@ -82,14 +60,6 @@ class FirebaseGameServicesGooglePlugin(private var activity: Activity? = null) :
             }
             return context.getString(stringRes)
         }
-
-        @JvmStatic
-        fun registerWith(registrar: PluginRegistry.Registrar) {
-            val channel = MethodChannel(registrar.messenger(), CHANNEL_NAME)
-            val plugin = FirebaseGameServicesGooglePlugin(registrar.activity())
-            channel.setMethodCallHandler(plugin)
-            registrar.addActivityResultListener(plugin)
-        }
     }
 
     private fun silentSignIn() {
@@ -98,14 +68,15 @@ class FirebaseGameServicesGooglePlugin(private var activity: Activity? = null) :
         val authCode = clientId ?: getResourceFromContext(context, "default_web_client_id")
 
         val builder = GoogleSignInOptions.Builder(
-            GoogleSignInOptions.DEFAULT_GAMES_SIGN_IN).requestServerAuthCode(authCode).requestScopes(Drive.SCOPE_APPFOLDER)
-
+            GoogleSignInOptions.DEFAULT_GAMES_SIGN_IN
+        ).requestServerAuthCode(authCode)
         googleSignInClient = GoogleSignIn.getClient(activity, builder.build())
+
         googleSignInClient?.silentSignIn()?.addOnCompleteListener { task ->
             pendingOperation = PendingOperation(method!!, gResult!!)
 
             if (task.isSuccessful) {
-                handleSignInResult(task.result)
+                handleSignInResult()
             } else {
                 Log.e("Error", "signInError", task.exception)
                 Log.i("ExplicitSignIn", "Trying explicit sign in")
@@ -119,34 +90,22 @@ class FirebaseGameServicesGooglePlugin(private var activity: Activity? = null) :
 
         val authCode = clientId ?: getResourceFromContext(context, "default_web_client_id")
 
-        val builder = GoogleSignInOptions.Builder(
-            GoogleSignInOptions.DEFAULT_GAMES_SIGN_IN
-        ).requestServerAuthCode(authCode).requestScopes(Drive.SCOPE_APPFOLDER)
-
-        googleSignInClient = GoogleSignIn.getClient(activity, builder.build())
+        googleSignInClient = GoogleSignIn.getClient(activity,
+            GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_GAMES_SIGN_IN)
+                .requestServerAuthCode(authCode).build())
         activity.startActivityForResult(googleSignInClient?.signInIntent, RC_SIGN_IN)
     }
 
-    private fun handleSignInResult(googleSignInAccount: GoogleSignInAccount) {
-        val activity = this.activity!!
+    private fun handleSignInResult() {
+        val activity = activity ?: return
 
-        val gamesClient =
-            Games.getGamesClient(activity, GoogleSignIn.getLastSignedInAccount(activity)!!)
+        achievementClient = PlayGames.getAchievementsClient(activity)
+        leaderboardsClient = PlayGames.getLeaderboardsClient(activity)
 
-
-        achievementClient = Games.getAchievementsClient(activity, googleSignInAccount)
-        leaderboardsClient = Games.getLeaderboardsClient(activity, googleSignInAccount)
-
-        gamesClient.setViewForPopups(activity.findViewById(android.R.id.content))
-        gamesClient.setGravityForPopups(Gravity.TOP or Gravity.CENTER_HORIZONTAL)
-
-        val account = GoogleSignIn.getLastSignedInAccount(activity)
-
-        if (account != null) {
-            if (method == Methods.signIn) {
-                signInFirebaseWithPlayGames(account)
-            } else if (method == Methods.signInLinkedUser) {
-                linkCredentialsFirebaseWithPlayGames(account)
+        GoogleSignIn.getLastSignedInAccount(activity)?.let { account ->
+            when (method) {
+                Methods.signIn -> signInFirebaseWithPlayGames(account)
+                Methods.signInLinkedUser -> linkCredentialsFirebaseWithPlayGames(account)
             }
         }
     }
@@ -155,9 +114,7 @@ class FirebaseGameServicesGooglePlugin(private var activity: Activity? = null) :
         val auth = FirebaseAuth.getInstance()
 
         val authCode = acct.serverAuthCode ?: throw Exception("auth_code_null")
-
         val credential = PlayGamesAuthProvider.getCredential(authCode)
-
         auth.signInWithCredential(credential).addOnCompleteListener { result ->
             if (result.isSuccessful) {
                 finishPendingOperationWithSuccess()
@@ -174,11 +131,8 @@ class FirebaseGameServicesGooglePlugin(private var activity: Activity? = null) :
         val auth = FirebaseAuth.getInstance()
 
         val currentUser = auth.currentUser ?: throw  Exception("current_user_null")
-
         val authCode = acct.serverAuthCode ?: throw Exception("auth_code_null")
-
         val credential = PlayGamesAuthProvider.getCredential(authCode)
-
 
         currentUser.linkWithCredential(credential).addOnCompleteListener { result ->
             if (result.isSuccessful) {
@@ -203,13 +157,12 @@ class FirebaseGameServicesGooglePlugin(private var activity: Activity? = null) :
             }
         }
     }
-    //endregion
 
     //region Achievements & Leaderboards
     private fun showAchievements(result: Result) {
         showLoginErrorIfNotLoggedIn(result)
         achievementClient?.achievementsIntent?.addOnSuccessListener { intent ->
-            activity?.startActivityForResult(intent, 0)
+            activity?.startActivityForResult(intent, RC_ACHIEVEMENT_UI)
             result.success("success")
         }?.addOnFailureListener {
             result.error("error", "${it.message}", null)
@@ -218,7 +171,7 @@ class FirebaseGameServicesGooglePlugin(private var activity: Activity? = null) :
 
     private fun unlock(achievementID: String, result: Result) {
         showLoginErrorIfNotLoggedIn(result)
-            achievementClient?.unlockImmediate(achievementID)?.addOnSuccessListener {
+        achievementClient?.unlockImmediate(achievementID)?.addOnSuccessListener {
             result.success("success")
         }?.addOnFailureListener {
             result.error("error", it.localizedMessage, null)
@@ -229,34 +182,40 @@ class FirebaseGameServicesGooglePlugin(private var activity: Activity? = null) :
         showLoginErrorIfNotLoggedIn(result)
         achievementClient?.incrementImmediate(achievementID, count)
             ?.addOnSuccessListener {
-            result.success("success")
-        }?.addOnFailureListener {
-            result.error("error", it.localizedMessage, null)
-        }
+                result.success("success")
+            }?.addOnFailureListener {
+                result.error("error", it.localizedMessage, null)
+            }
     }
 
     private fun showLeaderboards(leaderboardID: String, result: Result) {
         showLoginErrorIfNotLoggedIn(result)
-        leaderboardsClient?.getLeaderboardIntent(leaderboardID)?.addOnSuccessListener { intent ->
-            activity?.startActivityForResult(intent, 0)
+        val onSuccessListener: ((Intent) -> Unit) = { intent ->
+            activity?.startActivityForResult(intent, RC_LEADERBOARD_UI)
             result.success("success")
-        }?.addOnFailureListener {
-            result.error("error", it.localizedMessage, null)
+        }
+        val onFailureListener: ((Exception) -> Unit) = {
+            result.error("error", "${it.message}", null)
+        }
+        if (leaderboardID.isEmpty()) {
+            leaderboardsClient?.allLeaderboardsIntent?.addOnSuccessListener(onSuccessListener)?.addOnFailureListener(onFailureListener)
+        } else {
+            leaderboardsClient?.getLeaderboardIntent(leaderboardID)?.addOnSuccessListener(onSuccessListener)?.addOnFailureListener(onFailureListener)
         }
     }
 
     private fun submitScore(leaderboardID: String, score: Int, result: Result) {
         showLoginErrorIfNotLoggedIn(result)
         leaderboardsClient?.submitScoreImmediate(leaderboardID, score.toLong())?.addOnSuccessListener {
-        result.success("success")
+            result.success("success")
         }?.addOnFailureListener {
-        result.error("error", it.localizedMessage, null)
+            result.error("error", it.localizedMessage, null)
         }
     }
 
     private fun showLoginErrorIfNotLoggedIn(result: Result) {
         if (achievementClient == null || leaderboardsClient == null) {
-        result.error("error", "Please make sure to call signIn() first", null)
+            result.error("error", "Please make sure to call signIn() first", null)
         }
     }
     //endregion
@@ -265,162 +224,33 @@ class FirebaseGameServicesGooglePlugin(private var activity: Activity? = null) :
     private fun getPlayerID(result: Result) {
         showLoginErrorIfNotLoggedIn(result)
         val activity = activity ?: return
-        val lastSignedInAccount = GoogleSignIn.getLastSignedInAccount(activity) ?: return
-        Games.getPlayersClient(activity, lastSignedInAccount)
-        .currentPlayerId.addOnSuccessListener {
-            result.success(it)
-        }.addOnFailureListener {
-            result.error("error", it.localizedMessage, null)
-        }
+
+        PlayGames.getPlayersClient(activity)
+            .currentPlayer.addOnCompleteListener { task: Task<Player?> ->
+                result.success(task.result?.playerId)
+            }.addOnFailureListener { exception: Exception ->
+                result.error("error", exception.localizedMessage, null)
+            }
     }
 
     private fun getPlayerName(result: Result) {
         showLoginErrorIfNotLoggedIn(result)
         val activity = activity ?: return
-        val lastSignedInAccount = GoogleSignIn.getLastSignedInAccount(activity) ?: return
-        Games.getPlayersClient(activity, lastSignedInAccount)
-        .currentPlayer
-        .addOnSuccessListener { player ->
-            result.success(player.displayName)
-        }.addOnFailureListener {
-            result.error("error", it.localizedMessage, null)
-        }
+
+        PlayGames.getPlayersClient(activity)
+            .currentPlayer
+            .addOnCompleteListener { task: Task<Player?> ->
+                result.success(task.result?.displayName)
+            }.addOnFailureListener { exception: Exception ->
+                result.error("error", exception.localizedMessage, null)
+            }
     }
     //endregion
-
-   //region Save game
-
-    private fun getSavedGames(result: Result) {
-        snapshotsClient?.load(true)
-        ?.addOnCompleteListener { task ->
-            val gson = Gson()
-            val data = task.result.get()
-            if (data == null) {
-            result.error(
-                PluginError.failedToGetSavedGames.errorCode(),
-                PluginError.failedToGetSavedGames.errorMessage(),
-                null
-            )
-            return@addOnCompleteListener
-            }
-            val items = data
-            .toList()
-            .map { SavedGame(it.uniqueName, it.lastModifiedTimestamp, it.deviceName) }
-
-            val string = gson.toJson(items) ?: ""
-            result.success(string)
-            data.release()
-        }
-    }
-
-    private fun saveGame(
-        data: String, desc: String, name: String, result: Result
-    ) {
-        val metadataChange = SnapshotMetadataChange.Builder()
-        .setDescription(desc)
-        .build()
-        snapshotsClient?.open(name, true)
-        ?.addOnCompleteListener { task ->
-            val snapshot = task.result.data
-            if (snapshot != null) {
-            // Set the data payload for the snapshot
-            snapshot.snapshotContents.writeBytes(data.toByteArray())
-
-            // Commit the operation
-            snapshotsClient?.commitAndClose(snapshot, metadataChange)
-                ?.addOnSuccessListener {
-                result.success(null)
-                }
-                ?.addOnFailureListener {
-                result.error(PluginError.failedToSaveGame.errorCode(), it.localizedMessage, null)
-                }
-            } else {
-            result.error(
-                PluginError.failedToSaveGame.errorCode(),
-                PluginError.failedToSaveGame.errorMessage(),
-                null
-            )
-            }
-        }
-    }
-
-    private fun deleteGame(name: String, result: Result) {
-        // Open the saved game using its name.
-        snapshotsClient?.open(name, false)
-        ?.addOnFailureListener {
-            result.error(
-            PluginError.failedToDeleteSavedGame.errorCode(),
-            it.localizedMessage ?: "",
-            null
-            )
-        }
-        ?.continueWith { snapshotOrConflict ->
-            val snapshot = snapshotOrConflict.result.data
-            if (snapshot?.metadata == null) {
-                result.error(
-                PluginError.failedToDeleteSavedGame.errorCode(),
-                PluginError.failedToDeleteSavedGame.errorMessage(),
-                null
-                )
-                return@continueWith
-            }
-            snapshotsClient?.delete(snapshot.metadata)
-                ?.addOnSuccessListener {
-                result.success(it)
-                }
-                ?.addOnFailureListener {
-                result.error(
-                    PluginError.failedToDeleteSavedGame.errorCode(),
-                    it.localizedMessage ?: "",
-                    null
-                )
-                }
-        }
-    }
-
-    private fun loadGame(name: String, result: Result) {
-        // Open the saved game using its name.
-        snapshotsClient?.open(name, false)
-        ?.addOnFailureListener {
-            result.error(
-            PluginError.failedToLoadGame.errorCode(),
-            it.localizedMessage ?: "",
-            null
-            )
-        }
-        ?.continueWith {
-            val snapshot = it.result.data
-
-            // Opening the snapshot was a success and any conflicts have been resolved.
-            try {
-            // Extract the raw data from the snapshot.
-            val value = snapshot?.snapshotContents?.readFully()
-            if (value != null) {
-                result.success(String(value))
-            } else {
-                result.error(
-                PluginError.failedToLoadGame.errorCode(),
-                PluginError.failedToLoadGame.errorMessage(),
-                null
-                )
-            }
-            } catch (e: Exception) {
-            result.error(
-                PluginError.failedToLoadGame.errorCode(),
-                e.localizedMessage ?: "",
-                null
-            )
-            }
-        }
-    }
-
-    //endregion
-
-    //region Events
 
     override fun onAttachedToEngine(binding: FlutterPlugin.FlutterPluginBinding) {
         setupChannel(binding.binaryMessenger)
         context = binding.applicationContext
+        PlayGamesSdk.initialize(context)
     }
 
     override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
@@ -487,11 +317,8 @@ class FirebaseGameServicesGooglePlugin(private var activity: Activity? = null) :
                 )
             }
             else -> {
-                pendingOperation!!.result.error(
-                    "error",
-                    exception.localizedMessage,
-                    null
-                )
+                pendingOperation!!.result.error("error", exception.localizedMessage, null)
+                pendingOperation = null
             }
         }
     }
@@ -499,20 +326,20 @@ class FirebaseGameServicesGooglePlugin(private var activity: Activity? = null) :
     //region ActivityResultListener
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?): Boolean {
         if (requestCode == RC_SIGN_IN) {
-        val result = data?.let { Auth.GoogleSignInApi.getSignInResultFromIntent(it) }
+            val result = data?.let { Auth.GoogleSignInApi.getSignInResultFromIntent(it) }
 
-        val signInAccount = result?.signInAccount
+            val signInAccount = result?.signInAccount
 
-        if (result?.isSuccess == true && signInAccount != null) {
-            handleSignInResult(signInAccount)
-        } else {
-            finishPendingOperationWithError(ApiException(result?.status ?: Status(0)))
-            var message = result?.status?.statusMessage ?: ""
-            if (message.isEmpty()) {
-            message = "Something went wrong " + result?.status
+            if (result?.isSuccess == true && signInAccount != null) {
+                handleSignInResult()
+            } else {
+                finishPendingOperationWithError(ApiException(result?.status ?: Status(0)))
+                var message = result?.status?.statusMessage ?: ""
+                if (message.isEmpty()) {
+                    message = "Something went wrong " + result?.status
+                }
             }
-        }
-        return true
+            return true
         }
         return false
     }
@@ -542,11 +369,8 @@ class FirebaseGameServicesGooglePlugin(private var activity: Activity? = null) :
             }
             Methods.signIn -> {
                 method = Methods.signIn
-
                 clientId = call.argument<String>("client_id")
-
                 gResult = result
-
                 silentSignIn()
             }
             Methods.signInLinkedUser -> {
@@ -563,25 +387,7 @@ class FirebaseGameServicesGooglePlugin(private var activity: Activity? = null) :
             Methods.getPlayerName -> {
                 getPlayerName(result)
             }
-            Methods.saveGame -> {
-            val data = call.argument<String>("data") ?: ""
-            val name = call.argument<String>("name") ?: ""
-            saveGame(data, name, name, result)
-            }
-            Methods.loadGame -> {
-                val name = call.argument<String>("name") ?: ""
-                loadGame(name, result)
-            }
-            Methods.getSavedGames -> {
-                getSavedGames(result)
-            }
-            Methods.deleteGame -> {
-                val name = call.argument<String>("name") ?: ""
-                deleteGame(name, result)
-            }
-            else -> {
-                result.notImplemented()
-            }
+            else -> result.notImplemented()
         }
     }
 }
@@ -597,15 +403,4 @@ object Methods {
         "signInLinkedUser"
     const val getPlayerID = "getPlayerID"
     const val getPlayerName = "getPlayerName"
-    const val saveGame = "saveGame"
-    const val loadGame = "loadGame"
-    const val getSavedGames = "getSavedGames"
-    const val deleteGame = "deleteGame"
 }
-
-
-data class SavedGame(
-  val name: String?,
-  val modificationDate: Long?,
-  val deviceName: String?
-)